requirements.txt
openai==0.28
Flask>=3.0.3
python-dotenv==1.0.1
alembic==1.13.2
psycopg2==2.9.9
pytest-flask==1.3.0